requirements.txt
streamlit
moviepy
openai-whisper
requests
gtts
numpy
pillow
static-ffmpeg